services/api-AI-support/requirements.txt
absl-py==2.3.1
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
easyocr==1.7.2
facenet-pytorch==2.6.0
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.10.0
grpcio==1.67.1
grpcio-tools==1.67.1
idna==3.11
ImageIO==2.37.2
jax==0.7.1
jaxlib==0.7.1
Jinja2==3.1.6
kiwisolver==1.4.9
lazy_loader==0.4
MarkupSafe==3.0.3
matplotlib==3.10.7
mediapipe==0.10.21
ml_dtypes==0.5.3
mpmath==1.3.0
networkx==3.5
ninja==1.13.0
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pillow==10.2.0
pika==1.3.2
protobuf==4.25.8
pyclipper==1.3.0.post6
pycparser==2.23
pyparsing==3.2.5
python-bidi==0.6.7
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
scikit-image==0.25.2
scipy==1.16.3
sentencepiece==0.2.1
setuptools==80.9.0
shapely==2.1.2
six==1.17.0
sounddevice==0.5.3
sympy==1.14.0
tifffile==2025.10.16
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
wheel==0.45.1
